feat: breaking change `aten::size` -> `tract_core_shape_of`
previously related dimension were computed at export time and could not be
changed anymore.
This explicitly go against NNEF spec that depracted this practice but since
we do at SONOS manipulate networks that need to be stored with symbolic
dimensions that can be realised at runtime we need this extra-feature.